# core/api/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateObjectRequest": {
        "properties": {
          "body": {
            "description": "The body of the object",
            "example": "This is the body of the object. Markdown syntax is supported here.",
            "type": "string"
          },
          "icon": {
            "$ref": "#/components/schemas/Icon"
          },
          "name": {
            "description": "The name of the object",
            "example": "My object",
            "type": "string"
          },
          "properties": {
            "description": "The properties to set on the object; see ListTypes or GetType endpoints for linked properties",
            "items": {
              "$ref": "#/components/schemas/PropertyLinkWithValue"
            },
            "type": "array",
            "uniqueItems": false
          },
          "template_id": {
            "description": "The id of the template to use",
            "example": "bafyreictrp3obmnf6dwejy5o4p7bderaaia4bdg2psxbfzf44yya5uutge",
            "type": "string"
          },
          "type_key": {
            "description": "The key of the type of object to create",
            "example": "page",
            "type": "string"
          }
        },
        "required": [
          "type_key"
        ],
        "type": "object"
      },
...
      "ObjectResponse": {
        "properties": {
          "object": {
            "$ref": "#/components/schemas/ObjectWithBody"
          }
        },
        "type": "object"
      },
...
      "PaginatedResponse-Object": {
        "properties": {
          "data": {
            "description": "The list of items in the current result set",
            "items": {
              "$ref": "#/components/schemas/Object"
            },
            "type": "array",
            "uniqueItems": false
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object"
      },
...
      "ServerError": {
        "properties": {
          "code": {
            "example": "internal_server_error",
            "type": "string"
          },
          "message": {
            "example": "Internal server error",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 500,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "UnauthorizedError": {
        "properties": {
          "code": {
            "example": "unauthorized",
            "type": "string"
          },
          "message": {
            "example": "Unauthorized",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 401,
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "code": {
            "example": "bad_request",
            "type": "string"
          },
          "message": {
            "example": "Bad request",
            "type": "string"
          },
          "object": {
            "example": "error",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerauth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@anytype.io",
      "name": "Anytype Support",
      "url": "https://anytype.io/contact"
    },
    "description": "This API enables seamless interaction with Anytype's resources - spaces, objects, properties, types, templates, and beyond.",
    "license": {
      "name": "Any Source Available License 1.0",
      "url": "https://github.com/anyproto/anytype-api/blob/main/LICENSE.md"
    },
    "termsOfService": "https://anytype.io/terms_of_use",
    "title": "Anytype API",
    "version": "2025-11-08"
  },
...
  "paths": {
...
    "/v1/spaces/{space_id}/objects": {
      "get": {
        "description": "Retrieves a paginated list of objects in the given space. The endpoint takes query parameters for pagination (offset and limit) and returns detailed data about each object including its ID, name, icon, type information, a snippet of the content (if applicable), layout, space ID, blocks and details. It is intended for building views where users can see all objects in a space at a glance.\nSupports dynamic filtering via query parameters (e.g., ?done=false, ?created_date[gte]=2024-01-01, ?tags[in]=urgent,important). For select/multi_select properties you can use either tag keys or tag IDs, for object properties use object IDs. See FilterCondition enum for available conditions.",
        "operationId": "list_objects",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space in which to list objects; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The number of items to skip before starting to collect the result set",
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "type": "integer"
            }
          },
          {
            "description": "The number of items to return",
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "maximum": 1000,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse-Object"
                }
              }
            },
            "description": "The list of objects in the specified space"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "List objects",
        "tags": [
          "Objects"
        ]
      },
      "post": {
        "description": "Creates a new object in the specified space using a JSON payload. The creation process is subject to rate limiting. The payload must include key details such as the object name, icon, description, body content (which may support Markdown), source URL (required for bookmark objects), template identifier, and the type_key (which is the non-unique identifier of the type of object to create). Post-creation, additional operations (like setting featured properties or fetching bookmark metadata) may occur. The endpoint then returns the full object data, ready for further interactions.",
        "operationId": "create_object",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space in which to create the object; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObjectRequest"
              }
            }
          },
          "description": "The object to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectResponse"
                }
              }
            },
            "description": "The created object"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Create object",
        "tags": [
          "Objects"
        ]
      }
    },
    "/v1/spaces/{space_id}/objects/batch": {
      "patch": {
        "description": "Updates multiple objects in a single request.",
        "operationId": "update_object_batch",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space in which to update the objects; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateObjectBatchRequest"
              }
            }
          },
          "description": "The objects to update",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObjectResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The updated objects"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Resource not found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Update objects in batch",
        "tags": [
          "Objects"
        ]
      },
      "post": {
        "description": "Creates multiple objects in a single request.",
        "operationId": "create_object_batch",
        "parameters": [
          {
            "description": "The version of the API to use",
            "in": "header",
            "name": "Anytype-Version",
            "required": true,
            "schema": {
              "default": "2025-11-08",
              "type": "string"
            }
          },
          {
            "description": "The ID of the space in which to create the objects; must be retrieved from ListSpaces endpoint",
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObjectBatchRequest"
              }
            }
          },
          "description": "The objects to create",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObjectResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "The created objects"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerauth": []
          }
        ],
        "summary": "Create objects in batch",
        "tags": [
          "Objects"
        ]
      }
    },
...
  },
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://127.0.0.1:31009"
    }
  ]
}